requirement.txt
streamlit
pyttsx3
python-docx